refactor(panels): 보기 좋도록 하나의 패키지로 묶음
