+++ src/main/java/kr/codesquad/UI/Panels/WinNumberPanel.java
@@ -1,30 +1,27 @@
-package kr.codesquad.UI;
+package kr.codesquad.UI.Panels;
 
 import kr.codesquad.InputManager.AwtInputHandler;
 import kr.codesquad.InputManager.UserInputHandler;
+import kr.codesquad.UI.Panels.CustomPanel;
 
 import java.awt.*;
 import java.awt.event.ActionListener;
-import java.util.List;
-// importing Java AWT class
 
+public class WinNumberPanel extends CustomPanel {
 
-// class AWTExample2 directly creates instance of Frame class
-class SingleInputPanel extends CustomPanel  {
-
-    UserInputHandler inputHandler;
+    private UserInputHandler inputHandler;
     // initializing using constructor
 
     private Label errLabel;
     private Label instLabel;
     private String instruction;
     private TextField txt;
+    private TextField bonus;
     private Button btn;
-    SingleInputPanel(String instruction, ActionListener act) {
+    public WinNumberPanel( ActionListener act) {
         super();
         setSize(400,360);
         setLayout(null);
-        this.instruction = instruction;
         setElements();
         btn.addActionListener(act);
         inputHandler = new AwtInputHandler(btn, txt, null, errLabel);
@@ -32,26 +29,31 @@ class SingleInputPanel extends CustomPanel  {
 
     private void setElements(){
         errLabel = new Label();
-        instLabel = new Label(instruction);
+        instLabel = new Label("당첨번호를 입력해 주세요.");
         txt = new TextField();
+        bonus = new TextField();
         btn = new Button("확인");
 
+
         errLabel.setBounds(0, 100, 400, 30);
-        instLabel.setBounds(0, 135, 400, 30);
-        txt.setBounds(80, 170, 240, 30);
+        instLabel.setBounds(100, 135, 200, 30);
+        txt.setBounds(80, 170, 200, 30);
+        bonus.setBounds(300, 170, 30, 30);
         btn.setBounds(125, 230, 150, 30);
+
 
         errLabel.setAlignment(Label.CENTER);
         instLabel.setAlignment(Label.CENTER);
 
         add(errLabel);
         add(instLabel);
+        add(bonus);
         add(txt);
         add(btn);
     }
 
     @Override
-    public UserInputHandler getUserInputHandler(){
-        return inputHandler;
+    public UserInputHandler getUserInputHandler() {
+        return new AwtInputHandler(btn, txt, bonus, errLabel);
     }
 }
